+++ extensions/ql-vscode/src/common/vscode/abstract-webview.ts
@@ -11,11 +11,7 @@ import { join } from "path";
 
 import { DisposableObject, DisposeHandler } from "../../pure/disposable-object";
 import { tmpDir } from "../../helpers";
-import {
-  getHtmlForWebview,
-  WebviewMessage,
-  WebviewView,
-} from "../../interface-utils";
+import { getHtmlForWebview, WebviewMessage, WebviewView } from "./webview-html";
 
 export type WebviewPanelConfig = {
   viewId: string;
